requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
psycopg[c]~=3.1.9
drf-spectacular~=0.26.2
Pillow~=9.5.0
